adding show_handles resource
