+++ tests/test_templates.py
@@ -1,50 +1,5 @@
-import os
-
 import pytest
-
-import voila.app
 from lxml import etree
-
-BASE_DIR = os.path.dirname(__file__)
-
-
-class VoilaTest(voila.app.Voila):
-    def listen(self):
-        pass  # the ioloop is taken care of by the pytest-tornado framework
-
-
-@pytest.fixture
-def voila_app(voila_args, voila_config):
-    voila_app = VoilaTest.instance()
-    voila_app.initialize(voila_args + ['--no-browser'])
-    voila_config(voila_app)
-    nbconvert_template_path = os.path.join(
-        BASE_DIR, '..', 'share/jupyter/voila/templates/gridstack/nbconvert_templates/')
-    template_path = os.path.join(
-        BASE_DIR, '..', 'share/jupyter/voila/templates/gridstack/templates')
-    voila_app.nbconvert_template_paths.insert(0, nbconvert_template_path)
-    voila_app.template_paths.insert(0, template_path)
-    voila_app.start()
-    yield voila_app
-    voila_app.stop()
-    voila_app.clear_instance()
-
-
-@pytest.fixture
-def app(voila_app):
-    return voila_app.app
-
-
-@pytest.fixture
-def voila_config():
-    return lambda app: None
-
-
-@pytest.fixture
-def voila_args():
-    nb_path = os.path.join(BASE_DIR, 'nb.ipynb')
-    return [nb_path, '--VoilaTest.config_file_paths=[]']
-
 
 @pytest.mark.gen_test
 def test_template_test(http_client, base_url):
@@ -81,3 +36,6 @@ def test_template_test(http_client, base_url):
     elem = tree.xpath("//*[text()='This is a hidden cell.']")
     assert not elem
 
+    # test absence of handles
+    elem = tree.xpath("//div[contains(@class, 'gridhandle')]")
+    assert not elem
